저장/수정/삭제 시 물어보기

+++ src/components/BoardItem.tsx
@@ -7,6 +7,7 @@ import {removeBoard} from "../store/reducers/boardReducer";
 import Title from "./ui/Title";
 import UserInformation from "./common/UserInformation";
 import {BsTrash} from "react-icons/bs";
+import {removeConfirm} from "../utils/comfirm";
 
 type Props = {
 	board: Board;
@@ -17,7 +18,9 @@ export default function BoardItem({ board, uid }: Props) {
 	const dispatch = useDispatch();
 	const user = useSelector((state: RootState) => state.users[board.uid]);
 	const handleRemove = () => {
-		dispatch(removeBoard({ id: board.bid }));
+		if(removeConfirm()) {
+			dispatch(removeBoard({ id: board.bid }));
+		}
 	}
 	return (
 		<li className="w-full flex items-center justify-between gap-4 hover:bg-zinc-50 p-2">
